moved client source code into client subfolder
