+++ client/src/navbar.tsx
