+++ config/services.json
@@ -1,1 +1,1 @@
-{"services":[{"id":"1","name":"ascsadasdcasd","uri":"casdcasdc","timeout":20,"categories":[],"type":"","uiProps":{"top":68,"left":7}},{"id":"2","name":"CRM","uri":"http://localhost:3000/apis/1","timeout":20,"categories":["cascdasdc"],"type":"CRM","uiProps":{"top":441,"left":712}},{"id":"3","name":"DataHub","uri":"http://localhost:3000/apis/2","timeout":20,"categories":[],"type":"DB","uiProps":{"top":76,"left":255}},{"id":"4","name":"Broker","uri":"http://localhost:3000/apis/1","timeout":20,"categories":[],"type":"API","uiProps":{"top":569,"left":472}},{"id":"5","name":"DataHub ","uri":"http://localhost:3000/apis/1","timeout":20,"categories":[],"type":"API","uiProps":{"top":267,"left":377}},{"id":"6","name":"Northgate","uri":"http://localhost:3000/apis/2","timeout":30,"categories":[],"type":"DB","uiProps":{"top":443,"left":220}},{"id":"7","name":"steve","uri":"http://localhost:3000/apis/2","timeout":20,"categories":[],"type":"API","uiProps":{"top":248,"left":582}}]}
+{"services":[{"id":"1","name":"ascsadasdcasd","uri":"casdcasdc","timeout":20,"categories":[],"type":"","uiProps":{"top":68,"left":7}},{"id":"2","name":"CRM","uri":"http://localhost:3000/apis/1","timeout":20,"categories":["cascdasdc"],"type":"CRM","uiProps":{"top":441,"left":712}},{"id":"3","name":"DataHub","uri":"http://localhost:3000/apis/2","timeout":20,"categories":[],"type":"DB","uiProps":{"top":66,"left":216}},{"id":"4","name":"Broker","uri":"http://localhost:3000/apis/1","timeout":20,"categories":[],"type":"API","uiProps":{"top":569,"left":472}},{"id":"5","name":"DataHub ","uri":"http://localhost:3000/apis/1","timeout":20,"categories":[],"type":"API","uiProps":{"top":267,"left":377}},{"id":"6","name":"Northgate","uri":"http://localhost:3000/apis/2","timeout":30,"categories":[],"type":"DB","uiProps":{"top":443,"left":220}},{"id":"7","name":"steve","uri":"http://localhost:3000/apis/2","timeout":20,"categories":[],"type":"API","uiProps":{"top":248,"left":582}}]}
