+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const HWP = require('html-webpack-plugin');
 module.exports = {
-  entry: './src/index.tsx',
+  entry: './client/src/index.tsx',
   output: {
     filename: 'main.js',
     path: path.resolve(__dirname, 'dist'),
@@ -26,8 +26,9 @@ module.exports = {
   resolve: {
     extensions: ['.ts', '.tsx', '.js'],
     alias: {
-      '@app': path.resolve(__dirname, 'src/'),
+      '@app': path.resolve(__dirname, 'client/src/'),
+      '@images': path.resolve(__dirname, 'images'),
     },
   },
-  plugins: [new HWP({template: './src/index.html'})],
+  plugins: [new HWP({template: './client/src/index.html'})],
 };
